requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
python-jose[cryptography]==3.4.0
email-validator==2.1.0.post1
pydantic>=1.10,<2.0
starlette==0.27.0


